requirements.txt
asgiref==3.8.1
attrs==24.3.0
cattrs==24.1.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==4.2.20
django-extensions==4.1
django-unfold==0.63.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jsii==1.106.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mexc-sdk @ file:///C:/Users/geras/PycharmProjects/BybitBot/mexc_sdk-1_0_0-py3-none-any.whl
pillow==11.2.1
publication==0.0.3
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
shell==1.0.1
six==1.17.0
sqlparse==0.5.3
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
vk==3.0
Werkzeug==3.1.3
zipp==3.21.0
